Update README.md to make clear assumption of input files being present
Running the workflow requires use of input files.  These are to be located in the data folder, and their retrieval is described in the `README.md` file there.  However, currently, those directions are out of date, as per #2 .   We will need to update directions on how to process and/or retrieve these files to run this workflow.